Move PipedagConfig from util folder to core and add to public interface
