+++ src/pydiverse/pipedag/__init__.py
@@ -1,9 +1,10 @@
-from pydiverse.pipedag.core import Flow, Stage, Task
+from pydiverse.pipedag.core import Flow, PipedagConfig, Stage, Task
 from pydiverse.pipedag.materialize import Blob, Table, materialize
 
 __all__ = [
     "Blob",
     "Flow",
+    "PipedagConfig",
     "Stage",
     "Table",
     "materialize",
