+++ src/pydiverse/pipedag/core/__init__.py
@@ -1,3 +1,4 @@
+from .config import PipedagConfig
 from .flow import Flow
 from .result import Result
 from .stage import Stage
@@ -5,7 +6,8 @@ from .task import Task
 
 __all__ = [
     "Flow",
+    "PipedagConfig",
+    "Result",
     "Stage",
     "Task",
-    "Result",
 ]
