+++ src/pydiverse/pipedag/util/__init__.py
@@ -1,4 +1,3 @@
-from .config import PipedagConfig, setup_structlog
 from .deep_map import deep_map
 from .deep_merge import deep_merge
 from .disposable import Disposable
