+++ tests/test_cache/test_local_table_cache.py
@@ -4,13 +4,12 @@ from pathlib import Path
 import filelock
 import pytest
 
-from pydiverse.pipedag import materialize, Table, Flow, Stage
+from pydiverse.pipedag import materialize, Table, Flow, Stage, PipedagConfig
 import sqlalchemy as sa
 import pandas as pd
 
 from pydiverse.pipedag.backend.table.sql import sa_select
 from pydiverse.pipedag.context import StageLockContext
-from pydiverse.pipedag.util import PipedagConfig
 
 
 def get_flow():
