+++ tests/test_flows/test_source_invalidation.py
@@ -5,7 +5,6 @@ import pandas as pd
 from pydiverse.pipedag import Flow, Stage, Table, materialize
 from pydiverse.pipedag.backend.table.util.pandas import adj_pandas_types
 from pydiverse.pipedag.context import StageLockContext
-from pydiverse.pipedag.util.config import PipedagConfig
 
 dfA_source = pd.DataFrame(
     {
